404 if the Repo can't find a document
With the introduction of a Repo for fetching documents, the Repo would
throw an error if params[:section_id] didn't match a placeholder file
instead of allowing the controller to respond with a 404. This commit
checks for the existence of a file first, which if false allows the
controller to respond with a 404.

+++ app/repositories/manuals_repository.rb
@@ -16,7 +16,9 @@ private
     section_id ||= 'index'
     path = "public/#{manual_id}/#{section_id}.json"
 
-    build_ostruct(JSON.parse(File.open(path).read))
+    if File.exists?(path)
+      build_ostruct(JSON.parse(File.open(path).read))
+    end
   end
 
   def fetch_from_api(manual_id, section_id)
